requirements.txt
face_recognition==1.3.0
opencv_contrib_python==4.6.0.66
opencv_contrib_python_headless==4.8.1.78
opencv_python==4.6.0.66
Pillow==10.1.0
Pillow==10.2.0
PyAutoGUI==0.9.53
pyttsx3==2.90
pywin32==305
tkVideo==0.1
